Heron Verfahren verbessert Abbruchfunktion ist nun abhängig von der Differenz von dem aktuellen mit dem vorherigen Durchlauf der Schleife

+++ Aufgabe 2/Eintraege_und_Funktionen.c
@@ -52,7 +52,7 @@ int main (void)
                 f7(z1, z2);
             break;
             case 8:
-                f8(z1);
+                f8((double)z1);
             break;
             case 9:
                 OK = false;
